Rename file defining ForkableExample and exclude it in RSpec backtraces
* Before, RSpec failures would look like:
  # ./spec/datadog/profiling/collectors/stack_spec.rb:798:in 'block (4 levels) in <top (required)>'
  # ./spec/spec_helper.rb:313:in 'block (2 levels) in <top (required)>'
  # ./spec/spec_helper.rb:207:in 'block (2 levels) in <top (required)>'
  # ./spec/support/execute_in_fork.rb:32:in 'ForkableExample#run'

  which is confusing because it looks like it runs in a fork (execute_in_fork.rb) but it does not.
  This is even more ambiguous on older Rubies which do not show the class in the backtrace:

  # ./spec/support/execute_in_fork.rb:32:in `run'
